requirements.txt
# Core
transformers>=4.40.0
peft>=0.10.0
torch>=2.2.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

# Training
axolotl>=0.4.0
datasets>=2.18.0

# Security
pysigma>=0.11.0
mitreattack-python>=3.0.0
cyclonedx-bom>=4.0.0

# Observability
wandb>=0.16.0

# API
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# CLI
typer>=0.12.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pyyaml>=6.0.0
